+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## 1.0.0
  * Updated to `@animoca/ethereum-contracts-nft_staking:4.0.0`.
- * Staking is separated in phase 1 and phase 2.
- * Phase 2 requires escrowing of REVV for staking.
+ * New version of `NftStakingV2` used as a parent for `DeltaTimeStaking` that requires escrow for cars and drivers tokens.
 
 ## 0.4.0
  * Added `NFTRepairCentre`, a contract to manage defunct NFTs.